+++ tests/CollectionAbstractTest.php
@@ -49,21 +49,21 @@ class CollectionAbstractTest extends TestCase
      */
     public function twoCollectionsAreNotTheSameValueIfTheyHaveNotTheSameAmountOfValues(): void
     {
-        $iterator1 = $this->createMock(\ArrayIterator::class);
+        $iterator1 = $this->createMock(ArrayIterator::class);
         $iterator1
             ->expects($this->once())
             ->method('count')
             ->willReturn(1);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
-        $iterator2 = $this->createMock(\ArrayIterator::class);
+        $iterator2 = $this->createMock(ArrayIterator::class);
         $iterator2
             ->expects($this->once())
             ->method('count')
             ->willReturn(2);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
-        /* @var $collection1 \DigipolisGent\Value\CollectionInterface */
+        /** @var \DigipolisGent\Value\CollectionInterface $collection1 */
         $this->assertFalse($collection1->sameValueAs($collection2));
     }
 
@@ -75,11 +75,11 @@ class CollectionAbstractTest extends TestCase
     public function twoCollectionsAreNotTheSameIfArrayKeysAreDifferent(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
-        $iterator1 = new \ArrayIterator([0 => $value1]);
+        $iterator1 = new ArrayIterator([0 => $value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([1 => $value2]);
+        $iterator2 = new ArrayIterator([1 => $value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertFalse($collection1->sameValueAs($collection2));
@@ -97,11 +97,11 @@ class CollectionAbstractTest extends TestCase
             ->expects($this->once())
             ->method('sameValueAs')
             ->willReturn(false);
-        $iterator1 = new \ArrayIterator([$value1]);
+        $iterator1 = new ArrayIterator([$value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([$value2]);
+        $iterator2 = new ArrayIterator([$value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertFalse($collection1->sameValueAs($collection2));
@@ -119,11 +119,11 @@ class CollectionAbstractTest extends TestCase
             ->expects($this->once())
             ->method('sameValueAs')
             ->willReturn(true);
-        $iterator1 = new \ArrayIterator([$value1]);
+        $iterator1 = new ArrayIterator([$value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([$value2]);
+        $iterator2 = new ArrayIterator([$value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertTrue($collection1->sameValueAs($collection2));
@@ -140,7 +140,7 @@ class CollectionAbstractTest extends TestCase
      * @return \DigipolisGent\Value\CollectionInterface|\PHPUnit\Framework\MockObject\MockObject
      *   The mocked CollectionAbstract.
      */
-    protected function createCollectionMock($name, \ArrayIterator $iterator): CollectionInterface
+    protected function createCollectionMock($name, ArrayIterator $iterator): CollectionInterface
     {
         $mock = $this
             ->getMockBuilder(CollectionAbstract::class)