+++ src/ValueAbstract.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\Value;
 
 /**
@@ -12,10 +14,20 @@ namespace DigipolisGent\Value;
 abstract class ValueAbstract implements ValueInterface
 {
     /**
-     * @inheritdoc
+     * Compare of another ValueInterface has the exact same type.
+     *
+     * Checks if two value objects have the same class name.
+     * - Not valid if one of them is a parent class.
+     * - Not valid if one of them is a child class.
+     *
+     * @param \DigipolisGent\Value\ValueInterface $object
+     *   The value object to check.
+     *
+     * @return bool
+     *   True if the value type are the same, false if not.
      */
-    public function sameValueTypeAs(ValueInterface $object)
+    protected function sameValueTypeAs(ValueInterface $object): bool
     {
-        return \get_class($this) === \get_class($object);
+        return get_class($this) === get_class($object);
     }
 }