+++ tests/CollectionAbstractTest.php
@@ -1,8 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace DigipolisGent\Test\Value;
 
+use ArrayIterator;
 use DigipolisGent\Value\CollectionAbstract;
+use DigipolisGent\Value\CollectionInterface;
 use DigipolisGent\Value\ValueInterface;
 use PHPUnit\Framework\TestCase;
 
@@ -16,58 +20,66 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * Test if the iterator method returns an ArrayIterator.
+     *
+     * @test
      */
-    public function testGetIterator()
+    public function getIteratorMethodReturnsArrayIterator(): void
     {
         $collection = $this->getMockForAbstractClass(CollectionAbstract::class);
-        $this->assertInstanceOf(\ArrayIterator::class, $collection->getIterator());
+        $this->assertInstanceOf(ArrayIterator::class, $collection->getIterator());
     }
 
     /**
      * Not the same if they are not of the same type.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotSameValueIfDifferentTypes()
+    public function twoCollectionsAreNotSameValueIfDifferentTypes(): void
     {
-        $collection1 = $this->createCollectionMock('Type1', new \ArrayIterator());
-        $collection2 = $this->createCollectionMock('Type2', new \ArrayIterator());
+        $collection1 = $this->createCollectionMock('Type1', new ArrayIterator());
+        $collection2 = $this->createCollectionMock('Type2', new ArrayIterator());
 
         $this->assertFalse($collection1->sameValueAs($collection2));
     }
 
     /**
      * Not the same if they have not the same amount of values.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameValueIfTheyHaveNotTheSameAmountOfValues()
+    public function twoCollectionsAreNotTheSameValueIfTheyHaveNotTheSameAmountOfValues(): void
     {
-        $iterator1 = $this->createMock(\ArrayIterator::class);
+        $iterator1 = $this->createMock(ArrayIterator::class);
         $iterator1
             ->expects($this->once())
             ->method('count')
             ->willReturn(1);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
-        $iterator2 = $this->createMock(\ArrayIterator::class);
+        $iterator2 = $this->createMock(ArrayIterator::class);
         $iterator2
             ->expects($this->once())
             ->method('count')
             ->willReturn(2);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
-        /* @var $collection1 \DigipolisGent\Value\CollectionInterface */
+        /** @var \DigipolisGent\Value\CollectionInterface $collection1 */
         $this->assertFalse($collection1->sameValueAs($collection2));
     }
 
     /**
      * Not the same if the collection item keys are not the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameIfArrayKeysAreDifferent()
+    public function twoCollectionsAreNotTheSameIfArrayKeysAreDifferent(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
-        $iterator1 = new \ArrayIterator([0 => $value1]);
+        $iterator1 = new ArrayIterator([0 => $value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([1 => $value2]);
+        $iterator2 = new ArrayIterator([1 => $value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertFalse($collection1->sameValueAs($collection2));
@@ -75,19 +87,21 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * Not the same if the values on the same index are not the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreNotTheSameIfHaveNotTheSameValues()
+    public function twoCollectionsAreNotTheSameIfHaveNotTheSameValues(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
         $value1
             ->expects($this->once())
             ->method('sameValueAs')
             ->willReturn(false);
-        $iterator1 = new \ArrayIterator([$value1]);
+        $iterator1 = new ArrayIterator([$value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([$value2]);
+        $iterator2 = new ArrayIterator([$value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertFalse($collection1->sameValueAs($collection2));
@@ -95,19 +109,21 @@ class CollectionAbstractTest extends TestCase
 
     /**
      * The same if the values on the same index the same.
+     *
+     * @test
      */
-    public function testTwoCollectionsAreTheSameIfHaveTheSameValues()
+    public function twoCollectionsAreTheSameIfHaveTheSameValues(): void
     {
         $value1 = $this->createMock(ValueInterface::class);
         $value1
             ->expects($this->once())
             ->method('sameValueAs')
             ->willReturn(true);
-        $iterator1 = new \ArrayIterator([$value1]);
+        $iterator1 = new ArrayIterator([$value1]);
         $collection1 = $this->createCollectionMock('CollectionType', $iterator1);
 
         $value2 = $this->createMock(ValueInterface::class);
-        $iterator2 = new \ArrayIterator([$value2]);
+        $iterator2 = new ArrayIterator([$value2]);
         $collection2 = $this->createCollectionMock('CollectionType', $iterator2);
 
         $this->assertTrue($collection1->sameValueAs($collection2));
@@ -124,7 +140,7 @@ class CollectionAbstractTest extends TestCase
      * @return \DigipolisGent\Value\CollectionInterface|\PHPUnit\Framework\MockObject\MockObject
      *   The mocked CollectionAbstract.
      */
-    protected function createCollectionMock($name, \ArrayIterator $iterator)
+    protected function createCollectionMock($name, ArrayIterator $iterator): CollectionInterface
     {
         $mock = $this
             ->getMockBuilder(CollectionAbstract::class)