+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All Notable changes to the `digipolisgent/value` package.
 
+## [Unreleased]
+
+### Changed
+
+- Bumped minimum PHP version to 7.2.
+- Made the code strict.
+
+### Removed
+
+- Removed ValueFromArrayInterface and ValueToArrayInterface: better
+  pattern is using Normalizers to transform data into ValueObjects and
+  Serializers to transform ValueObjects into arrays. See the
+  [Flanders Basicregisters package](https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/tree/develop/src/Normalizer/FromJson)
+  as example how to normalize data into value objects.
+
 ## [1.1.2]
 
 ### Changed