+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All Notable changes to the `digipolisgent/value` package.
 
-## [Unreleased]
+## [2.0.0]
 
 ### Changed
 
@@ -45,6 +45,7 @@ All Notable changes to the `digipolisgent/value` package.
 - Added the CollectionInterface.
 - Added the CollectionAbstract class.
 
+[2.0.0]: https://github.com/digipolisgent/php_package_dg-value/compare/1.1.2...2.0.0
 [1.1.2]: https://github.com/digipolisgent/php_package_dg-value/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/digipolisgent/php_package_dg-value/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/digipolisgent/php_package_dg-value/compare/1.0.0...1.1.0